chore(users): cleanup code by refactoring into components
