+++ app/users/Header.tsx
@@ -1,0 +1,16 @@
+const TableHeader = () => {
+  return (
+    <thead>
+      <tr>
+        <th>Avatar</th>
+        <th>Contact</th>
+        <th>Role</th>
+        <th>Theme</th>
+        <th>Language</th>
+        <th>Joined</th>
+      </tr>
+    </thead>
+  );
+};
+
+export default TableHeader;
